# swagger.yaml
swagger: '2.0'
info:
  description: Набор сервисов для учета на предприятии
  version: "1.0.0"
  title: RestApi

produces:
  - application/json
  - application/xml
  - application/text

basePath: "/api"

paths:
  /reports/formats:
    get:
      summary: Получить список форматов для отчетов
      operationId: main.formats
      produces:
        - application/json
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: object
              properties:
                name:
                  type: string
                value:
                  type: string
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка 

  /reports/{category}/{format_type}:
    get:
      summary: Получить отчет по указанной категории и формату
      operationId: main.get_report
      parameters:
        - name: category
          in: path
          required: true
          type: string
          description: Категория данных для отчета (ranges, groups, nomenclature, receipts)
        - name: format_type
          in: path
          required: true
          type: string
          description: Формат отчета (CSV, MD, JSON, XML, DOCX, XLSX)
      responses:
        200:
          description: Успешная генерация отчета
          schema:
            type: string
        400:
          description: Неверные параметры
        500:
          description: Ошибка на сервере
        default:
          description: Неизвестная ошибка
  /crud/filter/{category}:
    post:
      summary: Фильтрация данных по категории
      operationId: main.filter_data
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: category
          in: path
          required: true
          type: string
          description: Категория данных для фильтрации (ranges, groups, nomenclature, receipts)
        - in: body
          name: filter
          description: "JSON-словарь для фильтрации данных"
          required: true
          schema:
            $ref: '#/definitions/Filter'
      responses:
        200:
          description: Успешная фильтрация данных
          schema:
            type: string
        400:
          description: Неверные параметры
        500:
          description: Ошибка на сервере

  /crud/transactions:
    post:
      summary: Получить список транзакций
      operationId: main.transactions
      produces:
        - application/json
      responses:
        200:
          description: Список транзакций успешно получен
          schema:
            type: object
        404:
          description: Транзакции не найдены
        500:
          description: Ошибка на сервере

  /crud/turnover:
    post:
      summary: Получить отчет по обороту
      operationId: main.turnover
      produces:
        - application/json
      responses:
        200:
          description: Отчет по обороту успешно создан
          schema:
            type: object
        404:
          description: Обороты не найдены
        500:
          description: Ошибка на сервере
  
  /crud/dateblock:
    post:
      summary: Расчитать оборот с датой блокировки
      operationId: main.dateblock
      produces:
        - application/json
      responses:
        200:
          description: Дата блока успешно обновлена
          schema:
            type: object
        404:
          description: Транзакции не найдены
        500:
          description: Ошибка на сервере

  /reports/dateblock:
    get:
      summary: Получить дату блокировки
      operationId: main.get_dateblock
      produces:
        - application/json
      responses:
        200:
          description: Блокировка дат успешно получена
          schema:
            type: object
            properties:
              message:
                type: array
                items:
                  type: string
        404:
          description: Файл блокировки дат не найден
        500:
          description: Ошибка на сервере

  /reports/nomenclature:
    get:
      summary: Получить номенклатуру
      operationId: main.get_nomenclature
      parameters:
        - name: query
          in: query
          required: true
          type: string
          description: Параметры для фильтрации номенклатуры
      responses:
        200:
          description: Успешное получение номенклатуры
          schema:
            type: object
        404:
          description: Номенклатура не найдена
        500:
          description: Ошибка на сервере

  /crud/nomenclature:
    put:
      summary: Добавить номенклатуру
      operationId: main.add_nomenclature
      parameters:
        - name: query
          in: query
          required: false
          type: string
          description: Параметры для добавления номенклатуры
      responses:
        200:
          description: Номенклатура успешно добавлена
          schema:
            type: object
        404:
          description: Ошибка добавления номенклатуры
        500:
          description: Ошибка на сервере

    patch:
      summary: Обновить номенклатуру
      operationId: main.update_nomenclature
      consumes:
        - application/json
      parameters:
        - in: body
          name: nomenclature
          description: Данные для обновления номенклатуры
          required: true
          schema:
            type: object
      responses:
        200:
          description: Номенклатура успешно обновлена
          schema:
            type: object
        404:
          description: Ошибка обновления номенклатуры
        500:
          description: Ошибка на сервере

    delete:
      summary: Удалить номенклатуру
      operationId: main.delete_nomenclature
      consumes:
        - application/json
      parameters:
        - in: body
          name: nomenclature
          description: Данные для удаления номенклатуры
          required: true
          schema:
            type: object
      responses:
        200:
          description: Номенклатура успешно удалена
          schema:
            type: object
        404:
          description: Ошибка удаления номенклатуры
        500:
          description: Ошибка на сервере

  /reports/balance_list:
    get:
      summary: Получить список балансов
      operationId: main.get_balance_list
      parameters:
        - name: data1
          in: query
          required: true
          type: string
          description: Начальная дата ОСВ
        - name: data2
          in: query
          required: true
          type: string
          description: Конечная дата ОСВ
      produces:
        - application/json
      responses:
        200:
          description: Список балансов успешно получен
          schema:
            type: object
        404:
          description: Балансов не найдено
        500:
          description: Ошибка на сервере

  /crud/save_data:
    post:
      summary: Сохранить данные
      operationId: main.save_data
      produces:
        - application/json
      responses:
        200:
          description: Данные успешно сохранены
          schema:
            type: object
        500:
          description: Ошибка при сохранении данных

  /crud/load_data:
    post:
      summary: Загрузить данные
      operationId: main.load_data
      produces:
        - application/json
      responses:
        200:
          description: Данные успешно загружены
          schema:
            type: object
        500:
          description: Ошибка при загрузке данных

definitions:
  Filter:
    type: object
    properties:
      name:
        type: string
        description: "Имя фильтра"
      id:
        type: string
        description: "ID фильтра"
      type:
        type: integer
        enum:
          - 1  # EQUALS
          - 2  # LIKE
        description: "Тип фильтра: 1 для EQUALS, 2 для LIKE"
    required:
      - name
      - id
      - type
    example:
      name: "example_name"
      id: ""
      type: 1